Move CLI decorators to a separate module
