+++ src/greynoise/cli/subcommand.py
@@ -1,12 +1,9 @@
 """CLI subcommands."""
 
-import functools
-
 import click
 
-from greynoise.cli.formatter import FORMATTERS
+from greynoise.cli.decorator import echo_result, handle_exceptions, pass_api_client
 from greynoise.cli.parameter import ip_address_parameter, ip_addresses_parameter
-from greynoise.exceptions import RequestFailure
 from greynoise.util import CONFIG_FILE, save_config, validate_ip
 
 
@@ -21,74 +18,6 @@ class SubcommandNotImplemented(click.ClickException):
     def __init__(self, subcommand_name):
         message = "{!r} subcommand is not implemented yet.".format(subcommand_name)
         super(SubcommandNotImplemented, self).__init__(message)
-
-
-def echo_result(function):
-    """Decorator that prints subcommand results correctly formatted.
-
-    :param function: Subcommand that returns a result from the API.
-    :type function: callable
-    :returns: Wrapped function that prints subcommand results
-    :rtype: callable
-
-    """
-
-    @functools.wraps(function)
-    def wrapper(*args, **kwargs):
-        result = function(*args, **kwargs)
-        context = click.get_current_context()
-        params = context.params
-        output_format = params["output_format"]
-        formatter = FORMATTERS[output_format]
-        if isinstance(formatter, dict):
-            # For the text formatter, there's a separate formatter for each subcommand
-            formatter = formatter[context.info_name]
-
-        output = formatter(result, params.get("verbose", False)).strip("\n")
-        click.echo(output)
-
-    return wrapper
-
-
-def handle_exceptions(function):
-    """Print error and exit on API client exception.
-
-    :param function: Subcommand that returns a result from the API.
-    :type function: callable
-    :returns: Wrapped function that prints subcommand results
-    :rtype: callable
-
-    """
-
-    @functools.wraps(function)
-    def wrapper(*args, **kwargs):
-        try:
-            return function(*args, **kwargs)
-        except RequestFailure as exception:
-            body = exception.args[1]
-            click.echo("API error: {}".format(body["error"]))
-            click.get_current_context().exit(-1)
-
-    return wrapper
-
-
-def pass_api_client(function):
-    """Create API client form API key and pass it to subcommand.
-
-    :param function: Subcommand that returns a result from the API.
-    :type function: callable
-    :returns: Wrapped function that prints subcommand results
-    :rtype: callable
-
-    """
-
-    @functools.wraps(function)
-    def wrapper(*args, **kwargs):
-        context = click.get_current_context()
-        api_client = context.obj["api_client"]
-        return function(api_client, *args, **kwargs)
-
-    return wrapper
 
 
 @click.command()
